// apps/cli/src/sdk/internal/generated/transport.ts
export type CreateAdminRoleDtoPermissionsItem = typeof CreateAdminRoleDtoPermissionsItem[keyof typeof CreateAdminRoleDtoPermissionsItem];


export const CreateAdminRoleDtoPermissionsItem = {
  systemmanage: 'system.manage',
  tasksview: 'tasks.view',
  taskscreate: 'tasks.create',
  tasksaccept: 'tasks.accept',
  taskscomplete: 'tasks.complete',
  tasksreview: 'tasks.review',
  tasksclose: 'tasks.close',
  demoreset: 'demo.reset',
} as const;

export interface CreateAdminRoleDto {
  name: string;
  permissions?: CreateAdminRoleDtoPermissionsItem[];
}

export interface CreateAdminUserDto {
  name: string;
  roleId: string;
}
// ...
export type UpdateAdminRoleDtoPermissionsItem = typeof UpdateAdminRoleDtoPermissionsItem[keyof typeof UpdateAdminRoleDtoPermissionsItem];


export const UpdateAdminRoleDtoPermissionsItem = {
  systemmanage: 'system.manage',
  tasksview: 'tasks.view',
  taskscreate: 'tasks.create',
  tasksaccept: 'tasks.accept',
  taskscomplete: 'tasks.complete',
  tasksreview: 'tasks.review',
  tasksclose: 'tasks.close',
  demoreset: 'demo.reset',
} as const;

export interface UpdateAdminRoleDto {
  name: string;
  permissions: UpdateAdminRoleDtoPermissionsItem[];
}

export interface UpdateAdminUserDto {
  name?: string;
  roleId?: string;
}